chore: configured application router with type safe routes
