+++ lib/config/config.dart
@@ -1,0 +1,2 @@
+export 'flavor.dart';
+export 'router/router.dart';
