+++ lib/config/router/router.dart
@@ -1,13 +1,2 @@
-import 'package:academia/config/router/app_navigation_observer.dart';
-import 'package:flutter/material.dart';
-import 'package:go_router/go_router.dart';
-
-class AppRouter {
-  static GlobalKey<NavigatorState> get globalNavigatorKey =>
-      GlobalKey<NavigatorState>();
-
-  static final router = GoRouter(
-    routes: [],
-    observers: [AppNavigationObserver()],
-  );
-}
+export 'app_router.dart';
+export 'routes.dart';
