+++ lib/config/router/routes.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+
+part 'routes.g.dart';
+
+@TypedGoRoute<HomeRoute>(path: "/")
+class HomeRoute extends GoRouteData with _$HomeRoute {
+  @override
+  Widget build(BuildContext context, GoRouterState state) {
+    return Scaffold(body: Center(child: Text("Hi")));
+  }
+}
